LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/last-successful-change-try-rule-expected.txt


PASS Starts rendering with --try
FAIL No successful position, keep --try assert_equals: expected 250 but got -50
PASS No successful position, last successful invalidated by @position-try change
